Added option for update and delete user.

+++ Restaurant.Database/Models/Table.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
 
-        public long Number { get; set; }
+        public int Number { get; set; }
 
         public bool IsTaken { get; set; }
