+++ Restaurant/ViewModels/CreateTableViewModel.cs
@@ -51,7 +51,7 @@ namespace Restaurant.ViewModels
             }
         }
 
-        public long Number { get; set; }
+        public int Number { get; set; }
 
         public int Seats
         {
